RBP - 23/05/2024 - Creado DTO y correcciones varias

+++ src/main/java/com/example/notiumb/dto/ListaOcioDTO.java
@@ -20,7 +20,7 @@ public class ListaOcioDTO {
     @PositiveOrZero
     private Double precio;
     @PositiveOrZero
-    private Double total_invitaciones;
+    private Integer total_invitaciones;
     @Valid
     private RppDTO rppDTO;
     private String detalleLista;
